Program_and_Product_Validator/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
pymongo==4.4.0
python-dotenv==1.0.0
PyYAML==6.0
python-dateutil==2.8.2
rapidfuzz==3.13.0
